docker/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
